test/requirements.txt
bleak
scapy
requests
ipaddress
asyncio
manuf

rich

# LLM / Voice Agent
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
fastmcp
python-dotenv
